requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.3
fonttools==4.61.1
fsspec==2026.2.0
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
MouseInfo==0.1.3
mpmath==1.3.0
mss==10.1.0
networkx==3.6.1
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
polars==1.38.1
polars-runtime-32==1.38.1
psutil==7.2.2
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==2.0.1
pyobjc-core==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-Quartz==12.1
pyparsing==3.3.2
pyperclip==1.11.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.18
PyQt5-stubs==5.15.6.0
PyQt5_sip==12.17.2
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytweening==1.2.0
PyYAML==6.0.3
requests==2.32.5
rubicon-objc==0.5.3
scipy==1.17.0
six==1.17.0
sympy==1.14.0
torch==2.10.0
torchvision==0.25.0
typing_extensions==4.15.0
ultralytics==8.4.12
ultralytics-thop==2.0.18
urllib3==2.6.3
